Bump the lower bound on semigroups to 0.16.2.
This reduces the CPP load, and allows the dependency to be dropped
on recent GHCs without turning into a real rat's nest of CPP.

In theory this might cause user trouble, but that would require
someone to be firmly stuck on a version of `semigroups` that's over
three years old at the time of writing while also wanting to use a
currently-bleeding-edge `semigroupoids`. As `semigroups` has a wide
compatibility range, this is rather unlikely.